Separate variable for every resource. Prefix for properties (if we lock several resources with the same type their properties will be overwritten by last locked  resource.

+++ src/test/java/org/jenkins/plugins/lockableresources/LockableResourceTest.java
@@ -200,7 +200,7 @@ public class LockableResourceTest {
 	public void testGetQueueItemId() {
 		System.out.println("getQueueItemId");
 		int expResult = 0;
-		int result = instance.getQueueItemId();
+		long result = instance.getQueueItemId();
 		assertEquals(expResult, result);
 	}
